Fix formatting errors
The file I downloaded has a big block of comments at the end and at the
end of each line in the comment there's a space. This really messed with
the formatting because when you save in VS code it removes things like
trailing whitespace. So if I open the file and save after already
committing, it will show that as a difference. I decided to just commit
that difference in spaces.