Rename and add files for main stack
I renamed some of the existin main game files like the gameboard one
and the input one and then I added some new ones. This new set of files
is my main "stack" now which are basically the main files for the
program: a game script that has the game stuff (separate from input),
the user and solver scripts are for connecting the input to game and for
connecting bot to the game resectively, and finally the minesweeper file
is basically like a main.py to run the connected files.