Seperate replay menu and switch reset to regen
For the winning or losing messages at the end of a game, there's a
little menu of options that comes with it. What I did was separate the
replay menu from the messages and have it in its individual function.
Also after adding the new regen function, which completley regenerates a
new board instead of resetting the same one, I switched all uses of
reset board to regen board.