Solve list reference error
I was using [[False] * C] * R to make the completed matrix at first.
But then I realized that because of that, the nested lists were copies
of each other and that was 100% messing with the solver. After removing,
I stopped after the weird error where many squares are just unrevealed.