Add display mask func def to overwrite
Because I added a solver mask to separate the solver's changes into a
separate mask, I know how to use a different display_mask method.
Originally, the display_mask method was written under the Game class,
and my Solver class inherited it from the Game class. But now that the
solver will have a different mask to display from, I'm overwriting the
display_mask method to incorporate the use of the 2 masks.