Switch all alters on mask board to alter solver mask
Because this is the solver file, any edits on the mask is done to color
it because of a solver move. So switching all occurences of editing the
mask to editing the solver mask simply separates solver edits to a
separate layer that kinda just overlays edits on the mask. Fock yeh 😎.