Add generators that loop through all nodes
I came up with this amazing idea to use generators to go through all the
nodes in the grid instead of having to use a 2D loop through the indexes
every time I need to go through all the nodes. I made three versions:
one that simply yields the nodes, the yields the coord, and one that's
like an enumeration cause it yields both the node and its coord.