Implement lake scan using BFS and DSU
Implemented lake scan with BFS, and then I figured out how to implement
it with DSU. I still gotta look at it in-depth cause I feel like it may
not have been implemented the best and it could improve. I need to go
through and thoroughly calculate the time complexity, also need to make
it add the shorter tree to the larger tree with doing Union.