Make unsolved tiles show real mines left
For unsolved tiles (yellow on edge of chain), some mines (or zero)
are flagged but not all. When matching patterns, we have to consider the
mines left and not the true mine count. So by changing the number shown
to the mines left, it makes it easier for the player to visually match
the patterns.