Add event queue clearing and pump events
During solving, any events (keys and mouse clicks) that occur will
process as soon as we get back to the event loop. I added event.clears
after any "animations", solving and win/lose animation, so that any
events that happened during the animation will be ignored after.

I then addressed the "Minesweeper not responding" by pumping events.
The not responding happens when you fail to make a call to the
event queue for too long, so pumping allows you to wake it up and
prevent the system from thinking the game locked up. I also added a
event poll directly after to get the last event, so that the queue
doesn't fill up with pumped events.