Create pattern class
Made a pattern class to represent the minesweeper patterns.
It stores the pattern as well as the actions to take with that pattern.
It stores information like the numbers that the pattern contains, these
act as the starting points for matching the patterns.

The pattern matching algorithm will traverse the board and anytime it
comes across an incomplete tile with a value in a contains set, that
will be a starting point for matching a patttern.